Print real xapian data instead of mocked values
The implementation is not finished yet, it is mainly to demonstrate
how I imagine the architecture.

The `search` function works alright except for sorting the results
by their importance / match ratio. When you search "mock" on the
current https://apps.fedoraproject.org/packages site, you get
"mock" and "mock-core-configs" as the first two results. In this
implementation they are not even on the first page. It may be caused
by incomplete cache and these pacakges missing there - it needs to
be investigated.

The pagination is also not finished yet.

The current implementation of `search_one` is just a dummy not
even returning the exact match of what is searched for.